Iblock: change namespace https://github.com/notamedia/console-jedi/pull/20#discussion_r94809329
