+++ src/Application/Application.php
@@ -16,7 +16,7 @@ use Notamedia\ConsoleJedi\Cache\Command\ClearCommand;
 use Notamedia\ConsoleJedi\Environment\Command\InitCommand;
 use Notamedia\ConsoleJedi\Module\Command as Module;
 use Notamedia\ConsoleJedi\Search\Command\ReIndexCommand;
-use Notamedia\ConsoleJedi\Schema\Schema;
+use Notamedia\ConsoleJedi\Iblock\Schema;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
