+++ src/Iblock/Exception/ExportException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Notamedia\ConsoleJedi\Iblock\Exception;
+
+/**
+ * Class ExportException
+ * @package Notamedia\ConsoleJedi\Iblock\Exception
+ */
+class ExportException extends SchemaException
+{
+
+}
