+++ src/Iblock/Exception/ImportException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Notamedia\ConsoleJedi\Iblock\Exception;
+
+/**
+ * Class ImportException
+ * @package Notamedia\ConsoleJedi\Iblock\Exception
+ */
+class ImportException extends SchemaException
+{
+
+}
