+++ src/Iblock/Exception/SchemaException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Notamedia\ConsoleJedi\Iblock\Exception;
+
+/**
+ * Class SchemaException
+ * @package Notamedia\ConsoleJedi\Iblock\Exception
+ */
+class SchemaException extends \RuntimeException
+{
+
+}
