--- src/Schema/Exception/ExportException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Notamedia\ConsoleJedi\Schema\Exception;
-
-/**
- * Class ExportException
- * @package Notamedia\ConsoleJedi\Schema\Exception
- */
-class ExportException extends SchemaException
-{
-
-}
