--- src/Schema/Exception/ImportException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Notamedia\ConsoleJedi\Schema\Exception;
-
-/**
- * Class ImportException
- * @package Notamedia\ConsoleJedi\Schema\Exception
- */
-class ImportException extends SchemaException
-{
-
-}
