--- src/Schema/Exception/SchemaException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Notamedia\ConsoleJedi\Schema\Exception;
-
-/**
- * Class SchemaException
- * @package Notamedia\ConsoleJedi\Schema\Exception
- */
-class SchemaException extends \RuntimeException
-{
-
-}
